updated libs and tesing in docker
