+++ docker/scala-single-test
@@ -1,0 +1,20 @@
+FROM databricksruntime/standard:10.4-LTS
+
+ARG SPARK_VERSION=3.2.1
+ARG TEST_TO_RUN
+ENV MVN_TEST_CMD="""-Dsuites="$TEST_TO_RUN""""
+
+RUN apt update && apt-get install -y maven locales
+
+# Ensure that we always use UTF-8 and with American English locale
+RUN locale-gen en_US.UTF-8
+
+ENV LC_ALL=en_US.UTF-8
+ENV LANG=en_US.UTF-8
+ENV LANGUAGE=en_US:en
+
+WORKDIR /tmp
+ADD pom.xml scalastyle-config.xml /tmp/
+RUN mvn verify --fail-never
+COPY . /tmp
+ENTRYPOINT mvn $MVN_TEST_CMD clean test
